Update getComputerChoice and getHumanChoice to return integer mappings
Removed logic to convert number choices to their string representations from getComputerChoice.

Changed the mapping values so I could use a computation algorithm to determine the winner.

Added rock, paper and scissor integer mappings as contants.

getHumanChoice now converts the string choice to an integer. This is so the playRound function
does not have to convert the choices to an integer in order to do the conversions.

Added global variables to keep track of the human and pc scores.