Refactor logic to work with ui
Renamed getHumanChoice to getPlayerChoice for consistenty with naming.

Removed commented code and tester functions that are not needed.

Refactored getPlayerChoice so it works with buttons and not text input.

Made the computation logic of playRound a little easier to follow using
constants instead of literals.

Renamed displayWinner as displayRoundWinner.

Placed results-text into a div called results-text-container.
Renamed class results-text to victory-status text.
Added another <p> element before victory-status-text and gave it a class
of computer-choice-text. This is so we can see what the computer chose.

Edited styling rules to properly display the new div.