Finish connecting game logic to ui
Changed all the <p> elements to labels.

Added a span element inside the labels displaying the score.
This makes it easier to store and retrieve the score values by separting
them from the label text.

Renamed the score labels to player-score-label and computer-score-label.

Renamed results-text label to victory status label. Added label to
results-text-container and named it computer-choice-text to display the
computer's choice.

Added a label to display the final winner message and a reset button
that appears at the end of the game.

Adjusted style sheet according to class name changes.

Removed playGame function and moved most of that code into the
input-buttons event listener.

Added updateScore function to update the score labels every round.
renamed declareGameWinner to checkFinalWinner.

Added resetGamefunction and added listener to resetButton.

Removed console.log and replaced with adding textContent to appropriate
labels.